.faaspot/requirements.txt
pycrypto==2.6.1
python-box==3.1.1
ipaddress==1.0.18